+++ lib/widgets/intent.dart
@@ -14,11 +14,13 @@ class IntentActions {
 class IntentDataKeys {
   static const action = 'action';
   static const allowMultiple = 'allowMultiple';
+  static const brightness = 'brightness';
   static const filters = 'filters';
   static const mimeType = 'mimeType';
   static const page = 'page';
   static const query = 'query';
   static const safeMode = 'safeMode';
+  static const secureUris = 'secureUris';
   static const uri = 'uri';
   static const widgetId = 'widgetId';
 }